+++ NEWS.md
@@ -4,6 +4,7 @@
 
 ## New features
 
+* Diagnostic function to produce density plots with stratification (#)
 * Added attributes as a return to the run_mod function (#24)
 * exported `run_mod()` function (#22)
 * Function that runs jags with option of stratification included. (#14)